chore: remove redundant passthrough
